+++ crates/tui/src/view/common/modal.rs
@@ -153,7 +153,8 @@ where
             .borders(Borders::ALL)
             .border_style(styles.border)
             .border_type(styles.border_type)
-            .title(modal.title());
+            .title(modal.title())
+            .style(styles.normal);
         // Add one cell of X padding so text doesn't butt up against the border;
         // that would interfere with word-based selection
         let margin = Margin::new(1, 0);